fix: change request token code and add new model about GetToken
